cerina_foundry/requirements.txt
langgraph
langchain-openai
pydantic
python-dotenv
fastapi 
uvicorn 
langchain 
langchain-core 
openai 
tiktoken 
httpx
langgraph-checkpoint-sqlite
mcp
axios 
lucide-react
requests
grandalf